requirements.txt
Django==4.2.5
djangorestframework==3.14.0
numpy==1.24.3
tensorflow==2.17.0
scikit-learn==1.2.2
pandas==2.0.1
pillow==9.5.0
